gradio-demo-app/requirements.txt
torch
torchvision
numpy
PyYAML
tqdm
gradio
Pillow
